migliorati i vari seeder + rinominati alcuni campi per maggiore chiarezza +

+++ database/migrations/2024_11_25_152015_create_orders_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->string('nome_guest');
             $table->string('mail_guest');
             $table->string('telefono_guest');
-            $table->integer('prezzo_totale');
+            $table->float('prezzo_totale');
             $table->timestamps();
         });
     }
